wiremock-loader/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
redis==5.0.1
aioredis==2.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
structlog==23.2.0
prometheus-client==0.19.0
asyncio-mqtt==0.16.1
aiokafka==0.8.11
python-multipart==0.0.6
python-json-logger==2.0.7
tenacity==8.2.3
jsonschema==4.20.0
pyyaml==6.0.1